Adds nonce protection for staking/unstaking calls.
