+++ bittensor/core/extrinsics/asyncex/weights.py
@@ -9,47 +9,14 @@ import bittensor.utils.weight_utils as weight_utils
 from bittensor.core.settings import version_as_int
 from bittensor.utils import format_error_message
 from bittensor.utils.btlogging import logging
+from bittensor.core.extrinsics.utils import (
+    async_sign_and_send_with_nonce as sign_and_send_with_nonce,
+)
 
 if TYPE_CHECKING:
     from bittensor_wallet import Wallet
     from bittensor.core.async_subtensor import AsyncSubtensor
     from bittensor.utils.registration import torch
-    from scalecodec.types import GenericCall
-
-
-async def sign_and_send_with_nonce(
-    subtensor: "AsyncSubtensor",
-    call: "GenericCall",
-    wallet: "Wallet",
-    wait_for_inclusion: bool,
-    wait_for_finalization: bool,
-    period: Optional[int] = None,
-):
-    """
-    Signs an extrinsic call with the wallet hotkey, adding an optional era for period
-    """
-    next_nonce = await subtensor.substrate.get_account_next_index(
-        wallet.hotkey.ss58_address
-    )
-
-    extrinsic_data = {"call": call, "keypair": wallet.hotkey, "nonce": next_nonce}
-    if period is not None:
-        extrinsic_data["era"] = {"period": period}
-
-    extrinsic = await subtensor.substrate.create_signed_extrinsic(**extrinsic_data)
-    response = await subtensor.substrate.submit_extrinsic(
-        extrinsic=extrinsic,
-        wait_for_inclusion=wait_for_inclusion,
-        wait_for_finalization=wait_for_finalization,
-    )
-
-    if not wait_for_finalization and not wait_for_inclusion:
-        return True, None
-
-    if await response.is_success:
-        return True, None
-
-    return False, format_error_message(await response.error_message)
 
 
 async def _do_commit_weights(
